Issue 310 - Pagination Test Suite Expanded
